Refactor to single attestationService.ts file
